refactor: collapse path module into hl7v2
