+++ crates/hl7v2-path/Cargo.toml
@@ -13,7 +13,7 @@ keywords = ["hl7", "healthcare", "path", "field", "access"]
 categories = ["parser-implementations"]
 
 [dependencies]
-thiserror = { workspace = true }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false }
 
 [dev-dependencies]
 proptest = "1.6"
