+++ crates/hl7v2/Cargo.toml
@@ -48,10 +48,10 @@ experimental-guard = ["dep:hl7v2-guard"]
 hl7v2-model = { version = "1.2.0", path = "../hl7v2-model" }
 hl7v2-escape = { version = "1.2.0", path = "../hl7v2-escape" }
 hl7v2-mllp = { version = "1.2.0", path = "../hl7v2-mllp" }
-hl7v2-path = { version = "1.2.0", path = "../hl7v2-path" }
 hl7v2-query = { version = "1.2.0", path = "../hl7v2-query" }
 hl7v2-parser = { version = "1.2.0", path = "../hl7v2-parser" }
 hl7v2-writer = { version = "1.2.0", path = "../hl7v2-writer" }
+thiserror = { workspace = true }
 
 hl7v2-json = { version = "1.2.0", path = "../hl7v2-json", optional = true }
 hl7v2-normalize = { version = "1.2.0", path = "../hl7v2-normalize", optional = true }
